### 2016-08-22 #72 @lukas-rokka
R Client for CAMS Radiation Service
R client for solar radiation time series of high resolution (15 min temporal; 3-5 km spatial; and covers Europe, Africa, Middle East).  It contains an interface to the CAMS radiation service, and two convenience functions that retrieves and parses the data into a data frame in a single call.

```
Package: camsRad
Title: Client for CAMS Radiation Service
Version: 0.1.1
Authors@R: person("Lukas", "Lundström", email = "lukas.rokka@gmail.com", role = c("aut", "cre"))
Description: CAMS radiations service provides time series of global, direct, and
    diffuse irradiations on horizontal surface, and direct irradiation on normal
    plane for the actual weather conditions as well as for clear-sky conditions.
    The geographical coverage is the field-of-view of the Meteosat satellite, roughly
    speaking Europe, Africa, Atlantic Ocean, Middle East. The time coverage of data is
    from 2004-02-01 up to 2 days ago. Data are available with a time step ranging from
    15 min to 1 month.  For more details see
    <http://www.soda-pro.com/web-services/radiation/cams-radiation-service>
    httr,
    xml2,
    readr
    roxygen2,
    ncdf4,
```

https://github.com/lukas-rokka/camsRad

Anybody in need of high resolution solar data. Solar energy, climate change, weather data for simulation etc

No

  - [ ] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.
