### 2016-08-23 #72 @sckott
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)
- [x] **Automated tests**

---

@lukas-rokka  Thanks for your submission!  Looking for reviewers
