### 2016-08-29 3bdf2cd
updated examples, e.g. lon to lng
