### 2016-08-29 3c99869
encode to "multipart". "raw" don't work with older httr
