### 2016-08-29 befd585
update for travis
