### 2016-09-21 #72 @Emaasit
Hello Scott,
I have just returned from my trip overseas and straight into doing my PhD
qualifying exams (which I had not anticipated to be this soon). I thought I
would have time at the beginning of the Semester to review this package.
Unfortunately, I will not be able to do so.

I apologize for any inconveniences.

Best Regards,
-- Daniel

On Tue, Sep 20, 2016 at 3:00 PM, Scott Chamberlain <notifications@github.com
> wrote:

> @Emaasit <https://github.com/Emaasit> - hey there, it's been 28 days,
> please get your review in soon, thanks 😺 (ropensci-bot)
>
