### 2016-10-05 #72 @jhollist
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

First, really nice job! I like the design with the general API function and then specific functions for typical use cases.  As someone who is ignorant of this type of data and CAMS in particular, I felt like I was able to follow along fairly well (although I had to look up McClear as I first thought it was a new McDonald's menu item...).  In short, `camsRad`:

  - does what it claims
  - has an easy to understand functions and documentation
  - should be a useful addition for those looking for solar radiation data

I do have a few comments, detailed below, that should be addressed prior to acceptance.

#### Documentation and Packaging Guidelines
- Include a Statement of Need
- Include a CONTRIBUTING file
- Maintainer - This was suggested to include in the reviewer template, but I didn't see any rOpenSci examples of this being done.  Probably OK as is.
- Add date and remember to update version to remove .9001 in DESCRIPTION
- Add CONDUCT.md.  See [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-code-of-conduct)
- Add top-level documentation for ?camsRad.  See [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-documentation) and a [minimum example](http://r-pkgs.had.co.nz/man.html#man-packages)

I am pretty new to the testing game, so take these with a grain of salt as I may be missing something.  But I had to jump through some hoops to get tests to work correctly.
- With covr::package_coverage() I get 79.07% and the tests themselves look reasonable. However, to get `covr` to work with the tests I had to turn off skip_on_cran.  I think there are ways to get the testing services to work with travis, see [here](https://github.com/jimhester/covr/issues/77).
- While setting CAMS_USERNAME locally is fine for running the tests, I am not sure how you would need to do this (other than hardcoding your own username) to get the tests to run on travis.  Perhaps @sckott can weigh in...
- Other than these issues, tests look good.

#### Dependencies
I too am a huge fan of `readr`, `dplyr`, tibbles and pipes.  That being said, I personally think it is best to avoid their use within packages as they are unnecessary dependencies since you can accomplish the same tasks with base R, might make maintenance of `camsRad` a bit more challenging (i.e. you need to keep track of more packages), and slows down install/CI/etc.  Others reading this might very likely have a different opinion, but were this my package I would:

- remove `readr` as a dependency and rely on base functions for reading in files, lines, etc.
- remove `dplyr` as a dependency and rely on base functions for manipulating data frames.
- avoid the use of `%>%` as it can potentially complicate debugging and is another unneeded dependency.
- return a plain old data.frame instead of a tibble.

#### Authentication
I am also quite new web APIs and especially authentication, so again maybe a good place for @sckott to weigh in.  But might be nice to add a function, like [here](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html#authentication) to deal with the username in a more consistent way.  This would also let it be set once instead of for each function call.

#### Support for all formats
You are currently providing support (or are planning on it) for 3 different return formats (csv, json, and netcdf).  I would assume that one of the goals for this package to is to make it easier to work with the CAMS API, but users probably don't really care which format the API returns, just as long as it is in an R friendly format.  Would probably be a lot easier if you picked one and built the package around that instead of trying to support all of them.  This is not something that I think needs to be addressed prior to acceptance (unless you want to), but might be something to think about for future releases and certainly prior to release to CRAN.

#### Some Code Comments and other nitpicking
- Vignette: Don't use Sys.Date(), add date in manually as you want the date to accurately reflect when the vignette was written (not rendered) or when the CAMS service request were last checked.
- lines 8 and 9: format should yyyy-mm-dd instead of yyy-mm-dd
- lines 144 and 169: no need to print(df), df should suffice
- line 149: function name.  Should this be cams_get_radiation?  If so should also change
- Take a look at [MangoTheCat/goodpractice](https://github.com/MangoTheCat/goodpractice) results. I ran and got:

```

 ✖ write unit tests for all functions, and all package code in general. 0% of code lines
    are covered by test cases.

    R/cams_rad.r:53:NA
    R/cams_rad.r:54:NA
    R/cams_rad.r:55:NA
    R/cams_rad.r:56:NA
    R/cams_rad.r:57:NA

  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and
    developers are used it and it is easier to read your code for them if you use '<-'.>

    tests/testthat/test-cams_api.R:6:11
    tests/testthat/test-cams_get_radiation.R:6:11

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    inst/doc/CAMS_solar_data.R:8:1
    R/cams_rad.r:52:1
    R/cams_rad.r:59:1
    R/cams_rad.r:64:1
    R/cams_rad.r:66:1

  ✖ not import packages as a whole, as this can cause name clashes between the imported
    packages. Instead, import only the specific functions you need.
  ✖ fix this R CMD check NOTE: Malformed Description field: should contain one or more
    complete sentences.
  ✖ fix this R CMD check NOTE: cams_parse: no visible binding for global variable ‘X1’
    Undefined global functions or variables: X1
```

The unit test lines in goodpractice will be different with skip_on_cran() turned off.
