### 2016-10-05 #72 @sckott
Thanks @jhollist for the review!

@lukas-rokka continue conversation here ...
