### 2016-10-06 #72 @lukas-rokka
Thanks @jhollist for your thorough review. I'll be away now for two weeks now, so just doing a commenting answer to begin with.

#### Tests

* I had trouble getting `covr` working, so I just skipped it. I see if I get it working with your info.

* I set the CAMS_USERNAME in Travis and it works fine.  If I don't set it, the test are skipped. Kind of cheating as I can get a pass without actually running the tests. But it can be good to be able to push to github without running tests on Travis, as the API calls are expensive and therefore limited to 15 calls per day (running tests and examples consume 4 API calls)

#### Dependencies
Here I partly disagree. I think I can remove the `readr` dependencies, and returning a `data.frame` instead of a `tibble` is maybe preferable. But using `dplyr` and `%>%` makes, in my opinion, the code easier to read and therefore also easier to maintain. I would prefer to continue using them.

#### Authentication
I look into that

### Support for all formats
Well, I would personally prefer to use just netcdf. But then I would get a dependency to the `ncdf4` package (now it is under `suggests:`).

In the `cams_api` function I try to cover all functionalities (and formats) that are available, so that the users can build their own functions from that.

The more specific functions (`cams_get_X`) returns R data frames and the user don't need to think about the formats at all.

I'll do a more thorough answer in about 3 weeks when I've gone through your suggestions in more detail.
