### 2016-10-06 #72 @jhollist
@lukas-rokka some quick response.

1. I think the reasons you outline for not running tests via travis make perfect sense and allows you to toggle it on and off as desired.
2. We can agree to disagree on the dependencies.  Leaving in dplyr and %>% won't have a bearing on my acceptance of the package. So no worries there.
3. Again the support for one over the others was a suggestion.  How you have it currently, makes a lot of sense.

Lastly, I'll reiterate that I think this is a nice package and will be a good addition!
