+++ .travis.yml
@@ -14,3 +14,9 @@ notifications:
   email:
     on_success: change
     on_failure: change
+
+
+#r_github_packages:
+#  - jimhester/covr
+#after_success:
+#  - Rscript -e 'covr::codecov()'
