+++ man/camsRad.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/camsRad_doc.R
+\docType{package}
+\name{camsRad}
+\alias{camsRad}
+\alias{camsRad-package}
+\title{R client for CAMS radiation service}
+\description{
+CAMS radiation service provides time series of global, direct, and diffuse
+irradiations on horizontal surface, and direct irradiation on normal plane
+for the actual weather conditions as well as for clear-sky conditions. The
+geographical coverage is the field-of-view of the Meteosat satellite, roughly
+speaking Europe, Africa, Atlantic Ocean, Middle East (-66° to 66° in both
+latitudes and longitudes). The time coverage of data is from 2004-02-01 up to
+2 days ago. Data are available with a time step ranging from 15 min to 1
+month. Target audience are researchers, developers and consultants in need of
+high resolution solar radiations time series. For details about the data, see
+[CAMS radiation service](http://www.soda-pro.com/web-services/radiation/cams-radiation-service)
+}
+
