### 2016-10-31 fde462a
added tests on cams_get_mcclear() and auth, gathered into one file
