### 2016-10-31 acc66fc
doc updated with new authentication
