+++ man/cams_get_radition.Rd
@@ -1,43 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cams_rad.r
-\name{cams_get_radition}
-\alias{cams_get_radition}
-\title{Retrieve CAMS solar radiation data}
-\usage{
-cams_get_radition(username, lat, lng, date_begin, date_end,
-  time_step = "PT01H", alt = -999, verbose = FALSE)
-}
-\arguments{
-\item{username}{The email registered at soda-pro.com.}
-
-\item{lat}{The latitude}
-
-\item{lng}{The longitude}
-
-\item{date_begin}{Start date as 'yyy-mm-dd' string}
-
-\item{date_end}{End date as 'yyy-mm-dd' string}
-
-\item{time_step}{'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
-hourly, 'P01D' for daily, 'P01M' for monthly}
-
-\item{alt}{The altitude in meters, use -999 to let CAMS decide}
-
-\item{verbose}{TRUE for verbose output}
-}
-\value{
-A tibble (data frame) with requested solar data
-}
-\description{
-Retrieve CAMS solar radiation data
-}
-\examples{
-\dontrun{
-username <- "your@email.com" # An email registered at soda-pro.com
-df <- cams_get_radition(username, lat=60, lng=15,
-                        date_begin="2016-01-01", date_end="2016-01-15")
-print(df)
-}
-
-}
-
