+++ man/cams_set_user.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cams_auth.R
+\name{cams_set_user}
+\alias{cams_set_user}
+\title{Set username used for authentication by CAMS radiation service}
+\usage{
+cams_set_user(username)
+}
+\arguments{
+\item{username}{Email registered at soda-pro.com.}
+}
+\description{
+Set username used for authentication by CAMS radiation service
+}
+\examples{
+\dontrun{
+cams_set_user("your@email.com") # An email registered at soda-pro.com
+}
+
+}
+
