+++ man/cams_api.Rd
@@ -1,19 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cams_rad.r
+% Please edit documentation in R/cams_api.R
 \name{cams_api}
 \alias{cams_api}
 \title{API client for
   \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
   radiation service}}
 \usage{
-cams_api(username, lat, lng, date_begin, date_end, alt = -999,
-  time_step = "PT01H", time_ref = "UT", verbose = FALSE,
-  service = "get_cams_radiation", format = "application/csv",
-  filename = "")
+cams_api(lat, lng, date_begin, date_end, alt = -999, time_step = "PT01H",
+  time_ref = "UT", verbose = FALSE, service = "get_cams_radiation",
+  format = "application/csv", filename = "")
 }
 \arguments{
-\item{username}{The email registered at soda-pro.com.}
-
 \item{lat}{The latitude}
 
 \item{lng}{The longitude}

[…]
