### 2016-10-31 f9e8e7b
split into three files. new authentication method. removed cams_parse() dependency on readr::
