+++ R/cams_api.R
@@ -2,20 +2,7 @@
 #'   \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
 #'   radiation service}
 #'
-#' @param username The email registered at soda-pro.com.
-#' @param lat The latitude
-#' @param lng The longitude
-#' @param date_begin Start date as 'yyy-mm-dd' string
-#' @param date_end End date as 'yyy-mm-dd' string
-#' @param alt The altitude in meters, use -999 to let CAMS decide
-#' @param time_step 'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
-#'   hourly, 'P01D' for daily, 'P01M' for monthly
-#' @param time_ref 'UT' for universal time, 'TST' for true solar time
-#' @param verbose TRUE for verbose output
-#' @param service 'get_mcclear' for CAMS McClear data, get_cams_radiation for CAMS
-#'   radiation data
-#' @param format 'application/csv', 'application/json',
-#'   'application/x-netcdf' or 'text/csv'
+#' @inheritParams cams_get_radiation
 #' @param filename path to file on disk to write to. If empty, data is kept in memory.
 #'
 #' @return list(ok=TRUE/FALSE, response=response). If ok=TRUE, response is the
@@ -47,9 +34,10 @@
 #' @import dplyr
 #' @export
 
-cams_api <- function(username, lat, lng, date_begin, date_end,
+cams_api <- function(lat, lng, date_begin, date_end,
                      alt=-999, time_step="PT01H", time_ref="UT", verbose=FALSE,
                      service="get_cams_radiation", format='application/csv', filename="") {
+  username <- cams_get_user()
   body <- paste0(
     '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>',
     '<wps:Execute service="WPS" version="1.0.0" ',
@@ -132,79 +120,3 @@ cams_api <- function(username, lat, lng, date_begin, date_end,
   return(list(ok=(r$status_code==200), respone=r))
 }
 
-#' Retrieve CAMS solar radiation data
-#' @inheritParams cams_api
-#'
-#' @return A tibble (data frame) with requested solar data
-#'
-#' @examples \dontrun{
-#' username <- "your@email.com" # An email registered at soda-pro.com
-#' df <- cams_get_radition(username, lat=60, lng=15,
-#'                         date_begin="2016-01-01", date_end="2016-01-15")
-#' print(df)
-#' }
-#'
-#' @export
-
-cams_get_radition <- function(username, lat, lng, date_begin, date_end,
-                              time_step="PT01H", alt=-999, verbose=FALSE) {
-  r <- cams_api(username, lat, lng, date_begin, date_end, alt, time_step=time_step,
-                verbose=verbose, service="get_cams_radiation", format="application/csv")
-  if(r$ok==FALSE) {
-    stop(r$response, call. = FALSE)
-  }
-  df <- cams_parse(readr::read_file(r$respone$content))
-  return(df)
-}
-
-#' Retrieve McClear clear sky solar radiation data
-#' @inheritParams cams_api
-#'
-#' @return A tibble (data.frame) with requested solar data
-#'
-#' @examples \dontrun{
-#' username <- "your@email.com" # An email registered at soda-pro.com
-#' df <- cams_get_mcclear(username, lat=60, lng=15,
-#'                        date_begin="2016-01-01", date_end="2016-01-15")
-#' print(df)
-#' }
-#'
-#' @export
-#'
-cams_get_mcclear <- function(username, lat, lng, date_begin, date_end,
-                             time_step="PT01H", alt=-999, verbose=FALSE) {
[…]
