+++ DESCRIPTION
@@ -8,16 +8,15 @@ Maintainer: Lukas Lundström <lukas.rokka@gmail.com>
 Description: CAMS radiations service provides time series of global, direct, and
     diffuse irradiations on horizontal surface, and direct irradiation on normal
     plane for the actual weather conditions as well as for clear-sky conditions.
-    The geographical coverage is the field-of-view of the Meteosat satellite, roughly
-    speaking Europe, Africa, Atlantic Ocean, Middle East. The time coverage of data is
-    from 2004-02-01 up to 2 days ago. Data are available with a time step ranging from
-    15 min to 1 month.  For more details see
+    The geographical coverage is the field-of-view of the Meteosat satellite,
+    roughly speaking Europe, Africa, Atlantic Ocean, Middle East. The time coverage
+    of data is from 2004-02-01 up to 2 days ago. Data are available with a time step
+    ranging from 15 min to 1 month. For more details see
     <http://www.soda-pro.com/web-services/radiation/cams-radiation-service>
 License: MIT + file LICENSE
 Depends: R (>= 3.1.0)
 Imports:
     httr (>= 1.2.1),
-    dplyr (>= 0.5.0),
     xml2 (>= 1.0.0)
 Suggests:
     ncdf4,
