### 2016-10-31 e21c008
remove non-ASCII
