+++ man/camsRad.Rd
@@ -10,9 +10,9 @@ CAMS radiation service provides time series of global, direct, and diffuse
 irradiations on horizontal surface, and direct irradiation on normal plane
 for the actual weather conditions as well as for clear-sky conditions. The
 geographical coverage is the field-of-view of the Meteosat satellite, roughly
-speaking Europe, Africa, Atlantic Ocean, Middle East (-66° to 66° in both
-latitudes and longitudes). The time coverage of data is from 2004-02-01 up to
-2 days ago. Data are available with a time step ranging from 15 min to 1
-month.
+speaking Europe, Africa, Atlantic Ocean, Middle East (-66 to 66 degrees in
+both latitudes and longitudes). The time coverage of data is from 2004-02-01
+up to 2 days ago. Data are available with a time step ranging from 15 min to
+1 month.
 }
 
