### 2016-10-31 efc7325
update doc and removed %>% & mutates
