+++ R/cams_get.R
@@ -7,12 +7,7 @@
 #' @param alt The altitude in meters, use -999 to let CAMS decide
 #' @param time_step 'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
 #'   hourly, 'P01D' for daily, 'P01M' for monthly
-#' @param time_ref 'UT' for universal time, 'TST' for true solar time
 #' @param verbose TRUE for verbose output
-#' @param service 'get_mcclear' for CAMS McClear data, get_cams_radiation for CAMS
-#'   radiation data
-#' @param format 'application/csv', 'application/json',
-#'   'application/x-netcdf' or 'text/csv'
 #'
 #' @return A data frame with requested solar data
 #'
@@ -83,13 +78,13 @@ cams_parse <- function(content) {
   message(writeLines(readLines(content, n=j)))
 
   # get column names, change first column name to 'timestamp'
-  col_names <- unlist((readLines(content, n=j))[j] %>% strsplit(";"))
+  col_names <- unlist(strsplit((readLines(content, n=j))[j], ";"))
   col_names <- c("timestamp", col_names[2:(length(col_names))])
 
   # get data, parse first column as datetime, set column names
-  df <- read.delim(content, header=FALSE, sep=";", skip=j) %>%
-    mutate(V1=as.POSIXct(substr(V1,23,38), "UTC", "%Y-%m-%dT%H:%M")) %>%
-    stats::setNames(col_names)
+  df <- utils::read.delim(content, header=FALSE, sep=";", skip=j)
+  df[, 1] <- as.POSIXct(substr(df[, 1], 23, 38), "UTC", "%Y-%m-%dT%H:%M")
+  names(df) <- col_names
 
   return(df)
 }
