+++ vignettes/CAMS_solar_data.Rmd
@@ -77,8 +77,9 @@ nc <- nc_open(r$respone$content)
 names(nc$var)
 
 # Create data.frame with datetime and global horizontal irradiation
-df <- data.frame(timestamp = as.POSIXct(nc$dim$time$vals, "UTC", origin="1970-01-01"),
-                 GHI = ncvar_get(nc, "GHI"))
+df <- data.frame(
+  timestamp = as.POSIXct(nc$dim$time$vals, "UTC", origin="1970-01-01"),
+  GHI = ncvar_get(nc, "GHI"))
 df$timestamp <- df$timestamp-3600*24 # shift timestamp 24 hours backwards
 
 nc_close(nc) # close connection
