+++ man/cams_get_radiation.Rd
@@ -12,9 +12,9 @@ cams_get_radiation(lat, lng, date_begin, date_end, time_step = "PT01H",
 
 \item{lng}{The longitude}
 
-\item{date_begin}{Start date as 'yyy-mm-dd' string}
+\item{date_begin}{Start date as 'yyyy-mm-dd' string}
 
-\item{date_end}{End date as 'yyy-mm-dd' string}
+\item{date_end}{End date as 'yyyy-mm-dd' string}
 
 \item{time_step}{'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
 hourly, 'P01D' for daily, 'P01M' for monthly}
@@ -22,14 +22,6 @@ hourly, 'P01D' for daily, 'P01M' for monthly}
 \item{alt}{The altitude in meters, use -999 to let CAMS decide}
 
 \item{verbose}{TRUE for verbose output}
-
-\item{time_ref}{'UT' for universal time, 'TST' for true solar time}
-
-\item{service}{'get_mcclear' for CAMS McClear data, get_cams_radiation for CAMS
-radiation data}
-
-\item{format}{'application/csv', 'application/json',
-'application/x-netcdf' or 'text/csv'}
 }
 \value{
 A data frame with requested solar data
