+++ NAMESPACE
@@ -4,4 +4,3 @@ export(cams_api)
 export(cams_get_mcclear)
 export(cams_get_radiation)
 export(cams_set_user)
-import(dplyr)
