+++ .travis.yml
@@ -15,7 +15,7 @@ notifications:
     on_success: change
     on_failure: change
 
-r_github_packages:
-  - jimhester/covr
-after_success:
-  - Rscript -e 'covr::codecov()'
+#r_github_packages:
+#  - jimhester/covr
+#after_success:
+#  - Rscript -e 'covr::codecov()'
