### 2016-10-31 c37aca7
removed non-ASCII
