### 2016-10-31 b329c21
removed non-ascii
