+++ vignettes/CAMS_solar_data.Rmd
@@ -23,7 +23,7 @@ if(Sys.getenv("CAMS_USERNAME")=="") {
 ## CAMS radiation service
 The CAMS radiation service provides time series of global, direct, and diffuse Irradiations on horizontal surface, and direct Irradiation on normal plane for the actual weather conditions as well as for clear-sky conditions. The data can be accessed manually on the CAMS radiation service [site](http://www.soda-pro.com/web-services/radiation/cams-radiation-service). The service is part of the [Copernicus Atmosphere Monitoring Service](http://atmosphere.copernicus.eu/) (CAMS).
 
-* The geographical coverage is the field-of-view of the Meteosat satellite, roughly speaking Europe, Africa, Atlantic Ocean, Middle East (-66° to 66° in both latitudes and longitudes).
+* The geographical coverage is the field-of-view of the Meteosat satellite, roughly speaking Europe, Africa, Atlantic Ocean, Middle East (-66 to 66 degrees in both latitudes and longitudes).
 * The time coverage of data is from 2004-02-01 up to 2 days ago. Data are available with a time step ranging from 15 min to 1 month. The service allows time steps of 1 min as well, these will be interpolated values from the 15 min time steps (accounting for changes in clear-sky conditions). 
 * The horizontal resolution is the original pixel of the Meteosat images, 3-5 km.  The CAMS radiation service has currently (as of 2016-11-01) limited the amount of requests to 15 per day. This limit may evolve. 
 
