+++ DESCRIPTION
@@ -1,9 +1,10 @@
 Package: camsRad
 Type: Package
 Title: Client for CAMS Radiation Service
-Version: 0.1.1.9003
+Version: 0.2.1.9001
 Date: 2016-10-30
 Authors@R: person("Lukas", "Lundstrom", email = "lukas.rokka@gmail.com", role = c("aut", "cre"))
+Maintainer: Lukas Lundstrom <lukas.rokka@gmail.com>
 Description: CAMS radiations service provides time series of global, direct, and
     diffuse irradiations on horizontal surface, and direct irradiation on normal
     plane for the actual weather conditions as well as for clear-sky conditions.
