### 2016-10-31 b660a3a
added cran comments
