+++ cran-comments.md
@@ -1,0 +1,15 @@
+This is a new releas. [Onboarding candidate at](https://github.com/ropensci/onboarding/issues/72)
+
+## Test environments
+* local Windows 7 x64 install, R 3.3.1
+* ubuntu 12.04 (on travis-ci), R 3.3.1
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 0 note
+
+## Reverse dependencies
+
+This is a new release, so there are no reverse dependencies.
+
