+++ DESCRIPTION
@@ -11,8 +11,7 @@ Description: CAMS radiations service provides time series of global, direct, and
     The geographical coverage is the field-of-view of the Meteosat satellite,
     roughly speaking Europe, Africa, Atlantic Ocean, Middle East. The time coverage
     of data is from 2004-02-01 up to 2 days ago. Data are available with a time step
-    ranging from 15 min to 1 month. For more details see
-    <http://www.soda-pro.com/web-services/radiation/cams-radiation-service>
+    ranging from 15 min to 1 month.
 License: MIT + file LICENSE
 Depends: R (>= 3.1.0)
 Imports:
@@ -24,7 +23,8 @@ Suggests:
     knitr,
     testthat,
     rmarkdown
-URL: https://github.com/lukas-rokka/camsRad
+URL: https://github.com/lukas-rokka/camsRad,
+    http://www.soda-pro.com/web-services/radiation/cams-radiation-service
 BugReports: https://github.com/lukas-rokka/camsRad/issues
 LazyData: TRUE
 RoxygenNote: 5.0.1
