### 2016-11-01 0f3713a
added cran comments
