+++ README.Rmd
@@ -95,5 +95,3 @@ nc_close(nc)
 * This package and functions herein are provided as is, without any guarantee.
 * Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 * Please [report any issues or bugs](https://github.com/lukas-rokka/camsRad/issues).
-
-
