+++ .travis.yml
@@ -10,12 +10,7 @@ addons:
       - libnetcdf-dev
       - netcdf-bin
 
-#notifications:
-#  email:
-#    on_success: change
-#    on_failure: change
-
-#r_github_packages:
-#  - jimhester/covr
-#after_success:
-#  - Rscript -e 'covr::codecov()'
+r_github_packages:
+  - jimhester/covr
+after_success:
+  - Rscript -e 'covr::codecov()'
