### 2016-11-01 #72 @lukas-rokka
@jhollist, now finished with the revision. I've adopted almost all of your suggestions.

#### Documentation and Packaging Guidelines
All suggestions adopted.

#### Tests
got `covr` to work now, both on travis and locally. Some more testing added.

#### Dependencies
Removed all suggested dependencies. In retrospective, going from `readr` to plain base was not worth the hassle, while skipping mutate, tibbles and pipes was quite effortless.

#### Authentication
Applied the suggested authentication method.

#### Support for all formats
Still supporting all formats, because of reasons mentioned above.

#### Some Code Comments and other nitpicking
Think I fixed/adopted most if not all points mentioned here.

That's all!
