### 2016-11-04 #72 @jhollist
Sorry for the delay in getting back!

All of your edits sound good to me and I am happy to sign off!

Nice job!
