### 2016-11-04 #72 @sckott
thanks for your help on this @jhollist

@lukas-rokka A few comments:

* in all `httr::content` calls (e.g., https://github.com/lukas-rokka/camsRad/blob/master/R/cams_api.R#L103), set encoding explicitly, and parse to plain text, like `content(x, "text", encoding = "UTF-8")`, then since you're using `xml2` already read xml like `xml2::xml2::read_xml`
* curious about the `tryCatch` here https://github.com/lukas-rokka/camsRad/blob/master/R/cams_api.R#L88 - any httr verb call shouldn't fail itself, even if there is a status 400/500 series client or server error.
* can remove `Date` field in description, gets added later, e.g., see https://github.com/hadley/devtools/issues/1327#issuecomment-244590948
* please add more examples in the man files for the exported functions. there's only 1 for each of them
* in your vignette: use `r Sys.Date()`, like this https://raw.githubusercontent.com/ropenscilabs/geojsonlint/master/vignettes/geojsonlint_vignette.Rmd
* the parameters for exported fxns could use better descriptions. e,g `lat` just says _The latitude_, whereas could specify what form the latitude should take. also note the required and optional parameters. and what default values are, if any
