+++ DESCRIPTION
@@ -1,8 +1,7 @@
 Package: camsRad
 Type: Package
 Title: Client for CAMS Radiation Service
-Version: 0.2.1.9001
-Date: 2016-10-30
+Version: 0.3.0
 Authors@R: person("Lukas", "Lundstrom", email = "lukas.rokka@gmail.com", role = c("aut", "cre"))
 Maintainer: Lukas Lundstrom <lukas.rokka@gmail.com>
 Description: CAMS radiations service provides time series of global, direct, and
