### 2016-11-07 f7f7231
add minors
