### 2016-11-07 a9a6fcd
documentation & corrected spelling error
