+++ R/cams_auth.R
@@ -1,8 +1,8 @@
 #' Set username used for authentication by CAMS radiation service
 #'
-#' @param username Email registered at soda-pro.com.
+#' @param username Email registered at soda-pro.com. Required
 #' @examples \dontrun{
-#' cams_set_user("your@email.com") # An email registered at soda-pro.com
+#' # cams_set_user("your@email.com") # An email registered at soda-pro.com
 #' }
 #'
 #' @export
