### 2016-11-07 cc15a00
improved error handling
