### 2016-11-07 #72 @lukas-rokka
Thanks for the feedback @sckott

#### Points 1 & 2 (API):
I improved the error handling of `cams_api` according to your suggestions and the `httr`'s  [Best practices for writing an API package vignette](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html)

#### Points 3 & 5 (dates):
Well, here yours and @jhollist suggestions are conflicting ;-). So I take it as there is no consensus how to do. I had it like you suggested before, but changed to present form after @jhollist review.

I choose to follow @sckott advice as I usually forget to manually update this kind of date stamps. Even though I understand @jhollist concern that now the date stamp represent the build time, not necessary update of the vignette's content.

#### Points 4 & & (documentation):
Improved according to your suggestions
