### 2016-11-07 #72 @jhollist
I did not know that about the Date!  Got a few DESCRIPTIONS of my own to update!
