### 2016-11-07 #72 @sckott
sorry about that date disagreement @jhollist , should have read more thoroughly - I do think it's best practice now to remove `Date`, but I'm not the decider

all looks good @lukas-rokka thanks for the change.
