### 2016-11-07 #72 @sckott
@lukas-rokka Approved! Thanks for being part of the rOpenSci community

next steps:

-   Add the rOpenSci footer to the README
-   Transfer the repo to ropenscilabs - I've created a team in that org. account, and you should have received an email about, after you accept the invitation, go to Settings->Transfer Ownership
-   After transferring to ropenscilabs, travis builds should work automatically under the new github account. but you do need to update the link in the badge in your readme
-  Update all links to github repo to replace `lukas-rokka` with `ropenscilabs`
